Move patch_signal to conftest
