+++ freqtrade/tests/test_freqtradebot.py
@@ -19,7 +19,7 @@ from freqtrade.persistence import Trade
 from freqtrade.rpc import RPCMessageType
 from freqtrade.state import State
 from freqtrade.strategy.interface import SellCheckTuple, SellType
-from freqtrade.tests.conftest import (log_has, log_has_re, patch_edge,
+from freqtrade.tests.conftest import (log_has, log_has_re, patch_edge, patch_get_signal,
                                       patch_exchange, patch_wallet)
 from freqtrade.worker import Worker
 
@@ -58,15 +58,6 @@ def get_patched_worker(mocker, config) -> Worker:
     patch_freqtradebot(mocker, config)
     return Worker(args=None, config=config)
 
-
-def patch_get_signal(freqtrade: FreqtradeBot, value=(True, False)) -> None:
-    """
-    :param mocker: mocker to patch IStrategy class
-    :param value: which value IStrategy.get_signal() must return
-    :return: None
-    """
-    freqtrade.strategy.get_signal = lambda e, s, t: value
-    freqtrade.exchange.refresh_latest_ohlcv = lambda p: None
 
 
 def patch_RPCManager(mocker) -> MagicMock:
